// src/app/views/pages/report/balance-sheet/model/IBalanceSheet.ts
export class IBalanceSheet {
   nature?: string;
   head?: string;
   summeryHead?: string;
   transactional?: string;
   businessPartner?: string;
   organization?: string;
   department?: string;
   warehouse?: string;
   location?: string;
   debit?: number;
   credit?: number;
   balance?: number;
   docDate?: string;
   docDate2?: string;

   constructor() {}
}